Update not to run "download" and "before build" section for resume option.
We should not run those section when it is the resume mode,
because the resume mode is expected to run after users finish
"download" and "before build" process, fail on "build" process.

As related refactoring, updated args_dict not to set element
that has value: None.

Fixes https://github.com/sclorg/rpm-list-builder/issues/11