Travis CI: Test on CentOS as well
Note that I had to change deltarpm from false to 0 due to yum.

Also had to sed the FROM image in Dockerfile, because it is not
possible to set via ARG.

Also had to change C.utf-8 locale to en_US.utf-8 everywhere
because C.utf-8 is not yet available on CentOS 7.